Moved over clause delta translation into clause translator.
Except for the ast-rewriter. This is just the first step;
need to still decouple it from the main translator.
