+++ src/ast2ram/ClauseTranslator.cpp
@@ -16,6 +16,7 @@
 
 #include "ast2ram/ClauseTranslator.h"
 #include "Global.h"
+#include "LogStatement.h"
 #include "ast/Aggregator.h"
 #include "ast/Clause.h"
 #include "ast/Constant.h"
@@ -23,6 +24,7 @@
 #include "ast/NilConstant.h"
 #include "ast/NumericConstant.h"
 #include "ast/RecordInit.h"
+#include "ast/Relation.h"
 #include "ast/StringConstant.h"
 #include "ast/UnnamedVariable.h"
 #include "ast/analysis/Functor.h"
@@ -41,15 +43,20 @@
 #include "ram/Break.h"
 #include "ram/Conjunction.h"
 #include "ram/Constraint.h"
+#include "ram/DebugInfo.h"
 #include "ram/EmptinessCheck.h"
 #include "ram/Filter.h"
 #include "ram/FloatConstant.h"
+#include "ram/LogRelationTimer.h"
+#include "ram/LogSize.h"
+#include "ram/LogTimer.h"
 #include "ram/Negation.h"
 #include "ram/NestedIntrinsicOperator.h"
 #include "ram/Project.h"
 #include "ram/Query.h"
 #include "ram/Relation.h"
 #include "ram/Scan.h"
+#include "ram/Sequence.h"
 #include "ram/SignedConstant.h"
 #include "ram/TupleElement.h"
 #include "ram/UnpackRecord.h"
@@ -61,6 +68,88 @@
 #include <vector>
 
 namespace souffle::ast2ram {
+
+VecOwn<ram::Statement> ClauseTranslator::generateClauseVersions(const std::set<const ast::Relation*>& scc,
+        const ast::Clause* cl, const AstToRamTranslator& tr, const TranslatorContext* context,
+        SymbolTable& symbolTable) {
+    VecOwn<ram::Statement> clauseVersions;
+
+    // Create each version
+    int version = 0;
+    const auto& atoms = ast::getBodyLiterals<ast::Atom>(*cl);
+    for (size_t i = 0; i < atoms.size(); i++) {
+        const auto* atom = atoms[i];
+
+        // Only interested in atoms within the same SCC
+        if (!contains(scc, context->getAtomRelation(atom))) {
+            continue;
+        }
+
+        appendStmt(clauseVersions, generateClauseVersion(scc, cl, i, version, tr, context, symbolTable));
+
+        // increment version counter
+        version++;
+    }
+
+    // Check that the correct number of versions have been created
+    if (cl->getExecutionPlan() != nullptr) {
+        int maxVersion = -1;
+        for (auto const& cur : cl->getExecutionPlan()->getOrders()) {
+            maxVersion = std::max(cur.first, maxVersion);
+        }
+        assert(version > maxVersion && "missing clause versions");
+    }
+
+    return clauseVersions;
+}
+
+Own<ram::Statement> ClauseTranslator::generateClauseVersion(const std::set<const ast::Relation*>& scc,
+        const ast::Clause* cl, size_t deltaAtomIdx, size_t version, const AstToRamTranslator& tr,
+        const TranslatorContext* context, SymbolTable& symbolTable) {
+    const auto& atoms = ast::getBodyLiterals<ast::Atom>(*cl);
+
+    // Modify the processed rule to use delta relation and write to new relation
+    auto fixedClause = tr.createDeltaClause(cl, deltaAtomIdx);
+
+    // Replace wildcards with variables to reduce indices
+    nameUnnamedVariables(fixedClause.get());
+
+    // Add in negated deltas for later recursive relations to simulate prev construct
+    for (size_t j = deltaAtomIdx + 1; j < atoms.size(); j++) {
+        const auto* atomRelation = context->getAtomRelation(atoms[j]);
+        if (contains(scc, atomRelation)) {
+            auto deltaAtom = souffle::clone(ast::getBodyLiterals<ast::Atom>(*fixedClause)[j]);
+            deltaAtom->setQualifiedName(getDeltaRelationName(atomRelation->getQualifiedName()));
+            fixedClause->addToBody(mk<ast::Negation>(std::move(deltaAtom)));
+        }
+    }
+
+    // Translate the resultant clause as would be done normally
+    Own<ram::Statement> rule =
+            ClauseTranslator(*context, symbolTable).translateClause(*fixedClause, *cl, version);
+
+    // Add loging
+    if (Global::config().has("profile")) {
+        const std::string& relationName = toString(cl->getHead()->getQualifiedName());
+        const auto& srcLocation = cl->getSrcLoc();
+        const std::string clauseText = stringify(toString(*cl));
+        const std::string logTimerStatement =
+                LogStatement::tRecursiveRule(relationName, version, srcLocation, clauseText);
+        const std::string logSizeStatement =
+                LogStatement::nRecursiveRule(relationName, version, srcLocation, clauseText);
+        rule = mk<ram::LogRelationTimer>(
+                std::move(rule), logTimerStatement, getNewRelationName(cl->getHead()->getQualifiedName()));
+    }
+
+    // Add debug info
+    std::ostringstream ds;
+    ds << toString(*cl) << "\nin file ";
+    ds << cl->getSrcLoc();
+    rule = mk<ram::DebugInfo>(std::move(rule), ds.str());
+
+    // Add to loop body
+    return mk<ram::Sequence>(std::move(rule));
+}
 
 Own<ram::Statement> ClauseTranslator::translateClause(
         const ast::Clause& clause, const ast::Clause& originalClause, int version) {
